+++ src/vue/forms/WithdrawalForm.vue
@@ -4,7 +4,6 @@
       <template v-if="assets.length">
         <form
           @submit.prevent="isFormValid() && showConfirmation()"
-          id="withdrawal-form"
           novalidate
         >
           <div class="app__form-row withdrawal__form-row">